Safari-611.1.10/LayoutTests/fast/dom/Window/showModalDialog-mandatory-parameter.html

<!DOCTYPE html>
<script src="../../../resources/js-test-pre.js"></script>
<script>
description("Checks that the first parameter to window.showModalDialog() is mandatory");

shouldThrow("window.showModalDialog()", "'TypeError: Not enough arguments'");
</script>
<script src="../../../resources/js-test-post.js"></script>
