LayoutTests/fast/dom/Window/showModalDialog-mandatory-parameter-expected.txt

CONSOLE MESSAGE: Window 'showModalDialog' function is deprecated and will be removed soon.
Checks that the first parameter to window.showModalDialog() is mandatory

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.showModalDialog() threw exception TypeError: Not enough arguments.
PASS successfullyParsed is true

TEST COMPLETE
